Seperated account test extensions into its own file
